Make Spark TP2 the guarded 8.75 GiB shared-image profile
