--- runtime/profiles/glm53-flash-nvfp4-tp2/profile.json
@@ -1,189 +1,0 @@
-{
-  "schema": "sparkring-serving-profile/v1",
-  "status": "implemented",
-  "name": "glm53-flash-nvfp4-tp2-mtp3",
-  "selection": "explicit",
-  "model": {
-    "repository": "local-inference-lab/GLM-5.3-Flash-NVFP4",
-    "revision": "520de24",
-    "revision_kind": "recorded revision prefix",
-    "variant": "original NVFP4"
-  },
-  "runtime_requirements": "dependencies.json",
-  "transport": {
-    "name": "tp2-rocenante-adaptive",
-    "manifest": "../../transport_profiles/tp2-rocenante-adaptive/manifest.json",
-    "manifest_sha256": "eb03cfde826974811be3bfe5d88f36d9de105b73358f3eaa56b9ed44f19127c4",
-    "physical_cage": "p0",
-    "path_count": 2,
-    "local_hca_indices": [
-      0,
-      2
-    ],
-    "hca_inventory": [
-      "rocep1s0f0",
-      "rocep1s0f1",
-      "roceP2p1s0f0",
-      "roceP2p1s0f1"
-    ],
-    "all_reduce_max_bytes": 16777216,
-    "all_gather_shard_max_bytes": 16777216
-  },
-  "sparkcache": {
-    "enabled": false,
-    "status": "unsupported",
-    "reason": "This checkpoint, source composition, and transport have no combined SparkCache qualification."
-  },
-  "lifecycle": {
-    "automatic_restart": false,
-    "automatic_start": false,
-    "memory_guard_service": "sparkring-memory-guard",
-    "memory_guard_floor_bytes": 2147483648
-  },
-  "environment": {
-    "B12X_POLICY_MODE": "auto",
-    "CUDA_DEVICE_MAX_CONNECTIONS": "32",
-    "CUDA_DEVICE_ORDER": "PCI_BUS_ID",
-    "CUTE_DSL_ARCH": "sm_121a",
-    "FLASHINFER_CUDA_ARCH_LIST": "12.1f",
-    "FLASHINFER_DISABLE_VERSION_CHECK": "1",
-    "HF_HUB_OFFLINE": "1",
-    "NCCL_CUMEM_ENABLE": "0",
-    "NCCL_DEBUG": "WARN",
-    "NCCL_IB_DISABLE": "0",
-    "NCCL_IB_GID_INDEX": "3",
-    "NCCL_IB_SUBNET_AWARE_ROUTING": "0",
-    "NCCL_IGNORE_CPU_AFFINITY": "1",
-    "NCCL_NET": "IB",
-    "NCCL_NET_PLUGIN": "none",
-    "NCCL_P2P_LEVEL": "SYS",
-    "NCCL_PROTO": "LL,LL128,Simple",
-    "OMP_NUM_THREADS": "2",
-    "PYTORCH_CUDA_ALLOC_CONF": "expandable_segments:True",
-    "TORCH_USE_RTLD_GLOBAL": "1",
-    "TRANSFORMERS_OFFLINE": "1",
-    "VLLM_B12X_KDA_PREFILL_COALESCING": "1",
-    "VLLM_B12X_MOE_FP4_FORCE_A16": "0",
-    "VLLM_ENABLE_PCIE_ALLREDUCE": "0",
-    "VLLM_GDN_SPEC_DECODE_METADATA_FASTPATH": "1",
-    "VLLM_GLM53_MHC_PREFILL_SHARD": "1",
-    "VLLM_GLM53_KDA_GATE_SIDE_STREAM": "0",
-    "VLLM_GLM53_SPLIT_MAMBA_BLOCK_SIZE": "256",
-    "VLLM_GLM53_SPLIT_TARGET_BLOCK_SIZE": "2048",
-    "VLLM_LM_HEAD_A16": "1",
-    "VLLM_MEMORY_PROFILE_INCLUDE_ATTN": "1",
-    "VLLM_MEMORY_PROFILER_ESTIMATE_CUDAGRAPHS": "1",
-    "VLLM_MTP_NVFP4_LM_HEAD": "1",
-    "VLLM_MXFP8_LM_HEAD": "1",
-    "VLLM_NO_USAGE_STATS": "1",
-    "VLLM_PLUGINS": "b12x_loader",
-    "VLLM_WORKER_MULTIPROC_METHOD": "spawn",
-    "VLLM_DISABLED_KERNELS": "MarlinFP8ScaledMMLinearKernel",
-    "PYTHONDONTWRITEBYTECODE": "1",
-    "PYTHONPATH": "",
-    "NCCL_LOCAL_INFERENCE_PATH": "/opt/sparkring/nccl-pci/libnccl.so.2.30.7",
-    "VLLM_NCCL_SO_PATH": "/opt/sparkring/nccl-pci/libnccl.so.2.30.7",
-    "LD_PRELOAD": "/opt/sparkring/nccl-pci/libnccl.so.2.30.7",
-    "NCCL_IB_EXTENDED_IPV4_GIDS": "1",
-    "NCCL_IB_PRESERVE_PCI_DOMAIN": "1",
-    "NCCL_IB_ROUTE_DIAGNOSTICS": "1"
-  },
-  "vllm_args": [
-    "/models/target",
-    "--host",
-    "0.0.0.0",
-    "--tensor-parallel-size",
-    "2",
-    "--pipeline-parallel-size",
-    "1",
-    "--decode-context-parallel-size",
-    "1",
-    "--nnodes",
-    "2",
-    "--distributed-executor-backend",
-    "mp",
-    "--mamba-cache-mode",
-    "align",
-    "--enable-prefix-caching",
-    "--enable-chunked-prefill",
-    "--dtype",
-    "bfloat16",
-    "--attention-backend",
-    "B12X",
-    "--moe-backend",
-    "b12x",
-    "--no-enable-flashinfer-autotune",
-    "--reasoning-parser",
-    "glm45",
-    "--tool-call-parser",
-    "glm47",
-    "--enable-auto-tool-choice",
-    "--port",
-    "8000",
-    "--master-port",
-    "29500",
-    "--master-addr",
-    "${MASTER_ADDR}",
-    "--node-rank",
-    "${NODE_RANK}",
-    "--served-model-name",
-    "GLM-5.3-Flash-NVFP4",
-    "--max-model-len",
-    "262144",
-    "--max-num-seqs",
-    "8",
-    "--max-num-batched-tokens",
-    "8192",
-    "--prefill-schedule-interval",
-    "8",
-    "--prefix-cache-retention-interval",
-    "0",
-    "--recurrent-checkpoint-policy",
-    "aligned",
-    "--block-size",
-    "256",
-    "--kv-cache-memory-bytes",
-    "7247757312",
-    "--kv-cache-dtype",
-    "fp8",
-    "--mm-processor-cache-gb",
-    "0",
-    "--mm-encoder-tp-mode",
-    "data",
-    "--linear-backend",
-    "b12x",
-    "--quantization",
-    "modelopt_mixed",
-    "--kda-prefill-backend",
-    "b12x",
-    "--gdn-decode-kernel",
-    "b12x",
-    "--generation-config",
-    "auto",
-    "--load-format",
-    "b12x",
-    "--model-loader-extra-config",
-    "{\"allocation\":\"managed\"}",
-    "--gpu-memory-utilization",
-    "0.87",
-    "--speculative-config",
-    "{\"method\": \"mtp\", \"num_speculative_tokens\": 3, \"moe_backend\": \"humming\", \"attention_backend\": \"B12X\"}",
-    "--compilation-config",
-    "{\"mode\": 0, \"cudagraph_mode\": \"FULL_AND_PIECEWISE\", \"max_cudagraph_capture_size\": 32}",
-    "--limit-mm-per-prompt",
-    "{\"image\": 4, \"video\": 0}"
-  ],
-  "qualification": {
-    "source_deployment": "qualified for bounded C1/C8 text workloads and a semantic smoke check",
-    "shared_image": "research-only until its exact profile composition is tested",
-    "loader_selection": "The measured loader hardcoded managed allocation. The shared loader selects it with model-loader-extra-config; that source composition requires separate image qualification.",
-    "nccl_selection": "The prospective shared image uses source-verified NCCL 2.30.7 with dual-domain routing flags. Measured single-DAC evidence used NCCL 2.30.4; it does not qualify this binary or toolchain combination.",
-    "limitations": [
-      "Long-duration serving stability is not established.",
-      "Image/video accuracy is not established; image count is only a configured limit.",
-      "Both cables remained connected; only p0 carried measured traffic.",
-      "Matched source hashes across two different rank image IDs do not establish one universal image.",
-      "The reported KV pool is an allocator estimate, not a measured maximum workload."
-    ]
-  }
-}
