--- runtime/profiles/glm53-flash-nvfp4-tp2/rank.env.example
@@ -1,4 +1,0 @@
-# Resolve these three host-specific settings independently on each rank.
-VLLM_HOST_IP=<this-rank-fabric-address>
-NCCL_SOCKET_IFNAME=<direct-fabric-interface>
-GLOO_SOCKET_IFNAME=<direct-fabric-interface>
