+++ runtime/profiles/glm53-flash-spark-tp2/runtime.env.example
@@ -1,47 +1,4 @@
-# Replace angle-bracket site values. Docker env files do not expand variables.
-LD_PRELOAD=/opt/sparkring/nccl/libnccl.so.2
-NCCL_NET_PLUGIN=none
-NCCL_IB_SUBNET_AWARE_ROUTING=0
-NCCL_IB_MERGE_NICS=0
-VLLM_HOST_IP=<this-node-address>
-VLLM_NCCL_SO_PATH=/opt/sparkring/nccl/libnccl.so.2
-OMP_NUM_THREADS=1
-PYTORCH_CUDA_ALLOC_CONF=expandable_segments:True
-FLASHINFER_CUDA_ARCH_LIST=12.1f
-NCCL_NET=IB
-NCCL_ALGO=Ring
-NCCL_CROSS_NIC=1
-VLLM_ENABLE_PCIE_ALLREDUCE=0
-CUTE_DSL_ARCH=sm_121a
-TORCH_CUDA_ARCH_LIST=12.1a
-XDG_CACHE_HOME=/cache/jit
-GLOO_SOCKET_IFNAME=<management-interface>
-VLLM_ALLREDUCE_USE_FLASHINFER=0
-VLLM_B12X_MOE_FP4_FORCE_A16=0
-NCCL_MAX_NCHANNELS=16
-TORCH_USE_RTLD_GLOBAL=1
-NCCL_IB_DISABLE=0
-NCCL_IB_HCA=<verified-roce-hca-list>
-NCCL_CUMEM_ENABLE=0
-NCCL_IGNORE_CPU_AFFINITY=1
-VLLM_NO_USAGE_STATS=1
-NCCL_P2P_LEVEL=SYS
-VLLM_ALLREDUCE_USE_SYMM_MEM=0
-VLLM_GLM53_SPLIT_TARGET_BLOCK_SIZE=2048
-VLLM_GLM53_SPLIT_MAMBA_BLOCK_SIZE=256
-VLLM_FASTSAFETENSORS_QUEUE_SIZE=0
-TORCHINDUCTOR_CACHE_DIR=/cache/jit/torchinductor
-B12X_CUTE_COMPILE_CACHE_DIR=/cache/jit/b12x
-NCCL_IB_GID_INDEX=3
-NCCL_PROTO=LL,LL128,Simple
-NCCL_BUFFSIZE=2097152
-HF_HUB_OFFLINE=1
-TRANSFORMERS_OFFLINE=1
-CMAKE_CUDA_ARCHITECTURES=121
-TRITON_CACHE_DIR=/cache/jit/triton
-NCCL_DEBUG=WARN
-VLLM_CACHE_ROOT=/cache/jit/vllm
-NCCL_SOCKET_IFNAME=<management-interface>
-VLLM_ENABLE_CUDA_COMPATIBILITY=0
-VLLM_USAGE_SOURCE=production-docker-image
-VLLM_PLUGINS=
+# Resolve these three host-specific settings independently on each rank.
+VLLM_HOST_IP=<this-rank-fabric-address>
+NCCL_SOCKET_IFNAME=<direct-fabric-interface>
+GLOO_SOCKET_IFNAME=<direct-fabric-interface>
