+++ runtime/sparkring/source_image/test_profile_assets.py
@@ -61,7 +61,7 @@ class ProfileAssetTests(unittest.TestCase):
             self.assertEqual(conflict.read_bytes(), b"different")
 
     def test_profile_dispatch_retains_common_verification_boundary(self):
-        profile = "glm53-flash-nvfp4-tp2-mtp3"
+        profile = "glm53-flash-spark-tp2-mtp3"
         manifest = {"runtime_profiles": {profile: {}, "tp4-dcp1-mtp3-prefill": {}},
                     "warmup_argv": ["/opt/sparkring/bin/serve-with-warmup.py"]}
         self.assertEqual(serving_argv(manifest, profile, ["/models/target"], "/usr/bin/python3"),
@@ -70,6 +70,32 @@ class ProfileAssetTests(unittest.TestCase):
                          manifest["warmup_argv"] + ["serve"])
         with self.assertRaisesRegex(ValueError, "absent"):
             serving_argv(manifest, "unlisted-profile", [])
+        with self.assertRaisesRegex(ValueError, "absent"):
+            serving_argv(manifest, "glm53-flash-nvfp4-tp2-mtp3", [])
+
+    def test_canonical_spark_tp2_assets_are_bound_to_the_five_profile_lock(self):
+        source = Path(__file__).resolve().parent
+        lock = json.loads((source / "glm53-tp4-lock.json").read_bytes())
+        name = "glm53-flash-spark-tp2-mtp3"
+        self.assertEqual(len(lock["profiles"]), 5)
+        self.assertNotIn("glm53-flash-nvfp4-tp2-mtp3", lock["profiles"])
+        entry = lock["profiles"][name]
+        self.assertEqual(entry["installed_profile_path"],
+                         "/opt/sparkring/profiles/glm53-flash-spark-tp2/profile.json")
+        with tempfile.TemporaryDirectory() as directory:
+            image = Path(directory)
+            data, record = prepare_assets(source.parents[2], lock, lock["source_date_epoch"])
+            install_assets(data, record, lock, image)
+            witness = verify_assets(lock, image)[entry["transport_profile"]]
+            self.assertEqual(witness["manifest_sha256"], entry["transport_manifest_sha256"])
+            profile_path = image / entry["installed_profile_path"].lstrip("/")
+            profile = json.loads(profile_path.read_bytes())
+            self.assertEqual(profile["name"], name)
+            self.assertTrue(profile["model"]["repository"].endswith("-NVFP4-Spark"))
+            profile["vllm_args"][profile["vllm_args"].index("--kv-cache-memory-bytes") + 1] = "1"
+            profile_path.write_text(json.dumps(profile))
+            with self.assertRaisesRegex(ValueError, "asset differs"):
+                verify_assets(lock, image)
 
 
 if __name__ == "__main__":
