+++ runtime/sparkring/test_launch.py
@@ -14,9 +14,14 @@ class LaunchTests(unittest.TestCase):
         with tempfile.TemporaryDirectory() as folder:
             root = Path(folder)
             env = root / 'runtime.env'
-            env.write_text('VLLM_HOST_IP=example\nNCCL_IB_HCA=example\nNCCL_SOCKET_IFNAME=example\nGLOO_SOCKET_IFNAME=example\n')
+            env.write_text('VLLM_HOST_IP=example\nNCCL_SOCKET_IFNAME=example\nGLOO_SOCKET_IFNAME=example\n')
+            model, cache = root / 'model', root / 'cache'
+            model.mkdir()
+            cache.mkdir()
+            (model / 'config.json').write_text('{}')
             image = 'ghcr.io/fujitsupolycom/sparkring@sha256:' + 'a'*64
-            name, command = launch.render(1, 'master.example', root, root, env, image)
+            plan = launch.render(1, 'master.example', model, cache, env, image)
+            name, command = plan['name'], plan['command']
             self.assertEqual(command[:2], ['docker', 'create'])
             self.assertIn('--headless', command)
             self.assertIn('master.example', command)
@@ -24,10 +29,10 @@ class LaunchTests(unittest.TestCase):
             self.assertNotIn('${NODE_RANK}', command)
             self.assertEqual(name, 'sparkring-glm53-tp2-r1')
             with self.assertRaises(ValueError):
-                launch.render(1, 'master.example', root, root, env, 'latest')
+                launch.render(1, 'master.example', model, cache, env, 'latest')
             env.write_text('VLLM_HOST_IP=<replace>')
             with self.assertRaises(ValueError):
-                launch.render(1, 'master.example', root, root, env, image)
+                launch.render(1, 'master.example', model, cache, env, image)
 
 
 if __name__ == '__main__':
